+++ src/api/Tokens.ts
@@ -7,21 +7,32 @@ import Configuration from '../Configuration'
  * @ignore
  * not part of public api, don't include in generated docs
  */
-// const isNode = new Function(
-//   'try {return this===global;}catch(e){return false;}'
-// )
-function isNode() {
-  return typeof window === 'undefined';
-}
+const isNode = new Function(
+  'try {return this===global;}catch(e){return false;}'
+)
 
+// https://nextjs.org/docs/app/building-your-application/optimizing/instrumentation#importing-runtime-specific-code
+/**
+ * @ignore
+ * not part of public api, don't include in generated docs
+ */
+const isEdgeRuntime = process.env.NEXT_RUNTIME === 'edge'
+
+/**
+ * @ignore
+ * not part of public api, don't include in generated docs
+ */
+const isServer = isNode() || isEdgeRuntime
 class Tokens {
   private accessTokenCookieName = `.access-token`
   private impersonationTokenCookieName = '.impersonation-token'
   private refreshTokenCookieName = '.refresh-token'
+  private identityTokenCookieName = '.identity-token'
 
   private accessToken?: string = null
   private impersonationToken?: string = null
   private refreshToken?: string = null
+  private identityToken?: string = null
 
   /**
    * @ignore
@@ -37,6 +48,9 @@ class Tokens {
     this.RemoveRefreshToken = this.RemoveRefreshToken.bind(this)
     this.SetImpersonationToken = this.SetImpersonationToken.bind(this)
     this.SetRefreshToken = this.SetRefreshToken.bind(this)
+    this.RemoveIdentityToken = this.RemoveIdentityToken.bind(this)
+    this.GetIdentityToken = this.GetIdentityToken.bind(this)
+    this.SetIdentityToken = this.SetIdentityToken.bind(this)
     this._isTokenExpired = this._isTokenExpired.bind(this)
     this._tryRefreshToken = this._tryRefreshToken.bind(this)
   }
@@ -46,18 +60,18 @@ class Tokens {
    */
 
   public GetAccessToken(): string | undefined {
-    return isNode() ? this.accessToken : cookies.get(this.accessTokenCookieName)
+    return isServer ? this.accessToken : cookies.get(this.accessTokenCookieName)
   }
 
   public SetAccessToken(token: string): void {
     parseJwt(token) // check if token is valid
-    isNode()
+    isServer
       ? (this.accessToken = token)
       : cookies.set(this.accessTokenCookieName, token)
   }
 
   public RemoveAccessToken(): void {
-    isNode()
+    isServer
       ? (this.accessToken = '')
       : cookies.remove(this.accessTokenCookieName)
   }
@@ -67,20 +81,20 @@ class Tokens {
    */
 
   public GetImpersonationToken(): string | undefined {
-    return isNode()
+    return isServer
       ? this.impersonationToken
       : cookies.get(this.impersonationTokenCookieName)
   }
 
   public SetImpersonationToken(token: string): void {
     parseJwt(token) // check if token is valid
-    isNode()
+    isServer
       ? (this.impersonationToken = token)
       : cookies.set(this.impersonationTokenCookieName, token)
   }
 
   public RemoveImpersonationToken(): void {
-    isNode()
+    isServer
       ? (this.impersonationToken = null)
       : cookies.remove(this.impersonationTokenCookieName)
   }
@@ -90,21 +104,43 @@ class Tokens {
    */
 
   public GetRefreshToken(): string | undefined {
-    return isNode()
+    return isServer
       ? this.refreshToken
       : cookies.get(this.refreshTokenCookieName)
   }
 
   public SetRefreshToken(token: string): void {
-    isNode()
+    isServer
       ? (this.refreshToken = token)
       : cookies.set(this.refreshTokenCookieName, token)
   }
 
   public RemoveRefreshToken(): void {
-    isNode()
+    isServer
       ? (this.refreshToken = null)
       : cookies.remove(this.refreshTokenCookieName)
+  }
+
+  /**
+   * Manage Identity Tokens
+   */
+
+  public GetIdentityToken(): string | undefined {
+    return isServer
+      ? this.identityToken
+      : cookies.get(this.identityTokenCookieName)
+  }
+
+  public SetIdentityToken(token: string): void {
+    isServer
+      ? (this.identityToken = token)
+      : cookies.set(this.identityTokenCookieName, token)
+  }
+
+  public RemoveIdentityToken(): void {
+    isServer
+      ? (this.identityToken = null)
+      : cookies.remove(this.identityTokenCookieName)
   }
 
   /**