+++ src/models/ExtendedLineItem.ts
@@ -30,7 +30,10 @@ export interface ExtendedLineItem<TExtendedLineItemXp = any, TFromUserXp = any, 
     readonly SupplierID?: string
     InventoryRecordID?: string
     readonly PriceScheduleID?: string
+    readonly IsOnSale?: boolean
     readonly PriceOverridden?: boolean
     Specs?: LineItemSpec[]
+    readonly IncomingOrderID?: string
+    readonly OutgoingOrderID?: string
     xp?: TExtendedLineItemXp
 }