+++ src/models/LineItem.ts
@@ -27,7 +27,10 @@ export interface LineItem<TLineItemXp = any, TProductXp = any, TVariantXp = any,
     readonly SupplierID?: string
     InventoryRecordID?: string
     readonly PriceScheduleID?: string
+    readonly IsOnSale?: boolean
     readonly PriceOverridden?: boolean
     Specs?: LineItemSpec[]
+    readonly IncomingOrderID?: string
+    readonly OutgoingOrderID?: string
     xp?: TLineItemXp
 }